create a separate file for helper function create date list from string
